pyproject.toml
[project]
name = 'spess'
version = '0.0'
requires-python = '>= 3.13'

dependencies = [
    'requests~=2.28',
    'rich~=13.3',
]

[project.scripts]

spess = 'spess.__main__:main'

[project.optional-dependencies]

dev = [
    'mypy~=1.16',
    'pyhumps~=3.8',
    'types-requests~=2.28',
]

[build-system]
requires = ['setuptools >= 77.0.3']
build-backend = 'setuptools.build_meta'

[tool.mypy]
packages = ['spess', 'spessgen']
